+++ Organization/Domain/EntityType.cs
@@ -1,0 +1,12 @@
+namespace Rezilio.Modules.Organization.Domain;
+
+public enum EntityType
+{
+    OrganizationalUnit,
+    Location,
+    Customer,
+    Supplier,
+    KeyPerson,
+    ItSystem,
+    BusinessProcess
+}